refactor: Validator InputView 클래스로 이동
