+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -2,17 +2,37 @@ package racingcar.view
 
 import racingcar.util.Constants
 import racingcar.util.Messages
-import racingcar.util.Validator
 import java.io.BufferedReader
 import java.io.InputStreamReader
 
-object InputView {
+class InputView {
     private val br = BufferedReader(InputStreamReader(System.`in`))
+
+    fun validateCarName(input: List<String>) {
+        require(input.all { it.isNotEmpty() && it.length <= Constants.CONDITION_CAR_MAX_LENGTH }) {
+            IllegalArgumentException(Messages.MESSAGE_INVALID_CAR_NAME)
+        }
+        require(input.size == input.toSet().size) {
+            IllegalArgumentException(Messages.MESSAGE_DUPLICATE_CAR_NAME)
+        }
+    }
+
+    fun validateCount(count: String): Int {
+        require(count.toIntOrNull() != null) {
+            IllegalArgumentException(Messages.MESSAGE_RACE_COUNT_NOT_INTEGER)
+        }
+
+        require(count.toInt() >= Constants.CONDITION_CAR_MIN_COUNT) {
+            IllegalArgumentException(Messages.MESSAGE_RACE_COUNT_TOO_SMALL)
+        }
+
+        return count.toInt()
+    }
 
     fun readCars(): List<String> {
         println(Messages.MESSAGE_INPUT_CAR_NAMES)
         val carsInput = br.readLine().split(Constants.DELIMITER_CAR_INPUT).map { it.trim() }
-        Validator.validateCarName(carsInput)
+        validateCarName(carsInput)
 
         return carsInput
     }
